Add mDNS support in VMs created with quick-vm
Install the avahi-daemon. Installing the package also enables the
service. The default configuration is good enough to use the only
network interface that is available in the VM other than loopback.

Note that support for mDNS within the VM does not necesssarily mean that
the host will be able to find the VM with mDNS. Firewall rules may need
to be updated. On Fedora workstation, the default rules block mDNS
traffic from the "libvirt" zone.